feat(landing): redesign sections, shadcn refactor, Framer Motion scroll
Implement the Claude Design handoff and rework the landing app:

- Design: bento Features grid, editorial timeline for "How it works",
  redesigned card FAQ, density/scale pass. Notch wordmark kept 1:1 with
  production (camera-lens mask + glow).
- Refactor: split the ~1.4k-line index.tsx into per-section components
  (components/sections/*) + shared lib/. Set up Tailwind v4 utilities
  mapped to the existing oklch tokens via @theme, with the page reset
  moved into @layer base so utilities win on the new components.
- shadcn/ui: Button (CVA variants), Accordion (FAQ), Tabs (install
  switcher) under components/ui, plus components.json and cn() helper.
- Fonts self-hosted via @fontsource-variable (no third-party request).
- Framer Motion (motion/react): convert the hand-rolled rAF scroll
  director to useScroll + useTransform (copy recede, MacBook rise/zoom,
  rim fade); notch whisper triggers via useMotionValueEvent once scrolled
  well into the zoom; nav floating + shared-layoutId active pill; CLI
  "meanwhile" reveal. Reduced-motion preserved and gated behind mount to
  avoid an SSR hydration mismatch.
- Fix: scroll cue ("See it munkel") anchored to the viewport so it shows
  at the top and fades before the demo caption.
