+++ apps/landing/src/components/ui/button.tsx
@@ -1,0 +1,52 @@
+import type { ComponentProps } from 'react'
+import { Slot } from '@radix-ui/react-slot'
+import { cva, type VariantProps } from 'class-variance-authority'
+
+import { cn } from '@/lib/utils'
+
+// Variants mirror the original hand-written buttons 1:1:
+//   primary → .btn-primary (filled, soft brand-tinted glow)
+//   outline → .btn-outline (hairline border)
+//   ghost   → .icon-btn    (transparent, hover surface — nav icon buttons)
+//   nav     → .nav-cta     (muted pill link in the nav bar)
+const buttonVariants = cva(
+  'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-[var(--radius-md)] border border-transparent text-sm font-medium transition-[background,border-color,color] outline-none focus-visible:ring-2 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50',
+  {
+    variants: {
+      variant: {
+        primary:
+          'bg-primary text-primary-foreground shadow-[var(--btn-primary-glow)] hover:bg-[var(--btn-primary-hover)]',
+        outline: 'border-border text-foreground hover:bg-muted',
+        ghost: 'text-muted-foreground hover:bg-muted hover:text-foreground',
+        nav: 'gap-1.5 rounded-full text-muted-foreground hover:text-foreground',
+      },
+      size: {
+        default: 'h-11 px-5',
+        icon: 'size-8',
+        nav: 'px-2 py-1.5',
+      },
+    },
+    defaultVariants: {
+      variant: 'primary',
+      size: 'default',
+    },
+  },
+)
+
+type ButtonProps = ComponentProps<'button'> &
+  VariantProps<typeof buttonVariants> & {
+    asChild?: boolean
+  }
+
+function Button({ className, variant, size, asChild = false, ...props }: ButtonProps) {
+  const Comp = asChild ? Slot : 'button'
+  return (
+    <Comp
+      data-slot="button"
+      className={cn(buttonVariants({ variant, size }), className)}
+      {...props}
+    />
+  )
+}
+
+export { Button, buttonVariants }
