+++ apps/landing/src/components/ui/tabs.tsx
@@ -1,0 +1,36 @@
+import type { ComponentProps } from 'react'
+import * as TabsPrimitive from '@radix-ui/react-tabs'
+
+import { cn } from '@/lib/utils'
+
+function Tabs({ className, ...props }: ComponentProps<typeof TabsPrimitive.Root>) {
+  return <TabsPrimitive.Root data-slot="tabs" className={cn(className)} {...props} />
+}
+
+function TabsList({ className, ...props }: ComponentProps<typeof TabsPrimitive.List>) {
+  return (
+    <TabsPrimitive.List
+      data-slot="tabs-list"
+      className={cn('inline-flex items-center', className)}
+      {...props}
+    />
+  )
+}
+
+function TabsTrigger({ className, ...props }: ComponentProps<typeof TabsPrimitive.Trigger>) {
+  return (
+    <TabsPrimitive.Trigger
+      data-slot="tabs-trigger"
+      className={cn('cursor-pointer outline-none disabled:pointer-events-none', className)}
+      {...props}
+    />
+  )
+}
+
+function TabsContent({ className, ...props }: ComponentProps<typeof TabsPrimitive.Content>) {
+  return (
+    <TabsPrimitive.Content data-slot="tabs-content" className={cn(className)} {...props} />
+  )
+}
+
+export { Tabs, TabsList, TabsTrigger, TabsContent }
