+++ apps/landing/src/lib/constants.ts
@@ -1,0 +1,8 @@
+// Shared external links + the Homebrew one-liner, referenced across sections.
+export const GITHUB_URL = 'https://github.com/limehq/munkel'
+export const DOWNLOAD_URL = `${GITHUB_URL}/releases/latest`
+export const PROTOCOL_URL = `${GITHUB_URL}/blob/main/apps/server/src/protocol.ts`
+export const CLI_URL = `${GITHUB_URL}/tree/main/apps/cli`
+export const LICENSE_URL = `${GITHUB_URL}/blob/main/LICENSE`
+export const DISCUSSIONS_URL = `${GITHUB_URL}/discussions`
+export const BREW_CMD = 'brew install limehq/tap/munkel'
