+++ apps/landing/src/routes/index.tsx
@@ -1,1312 +1,36 @@
-import { useEffect, useRef, useState } from 'react'
 import { createFileRoute } from '@tanstack/react-router'
-import {
-  ArrowRight,
-  Bot,
-  Check,
-  ChevronDown,
-  Copy,
-  Download,
-  Info,
-  Laptop,
-  Lock,
-  MonitorOff,
-  MonitorSmartphone,
-  Moon,
-  Package,
-  Rocket,
-  Sun,
-  Terminal,
-  TimerOff,
-  UserRoundX,
-  Wifi,
-} from 'lucide-react'
+import { MotionConfig } from 'motion/react'
+
+import { AnnounceBar } from '@/components/sections/announce-bar'
+import { Nav } from '@/components/sections/nav'
+import { Hero } from '@/components/sections/hero'
+import { HowItWorks } from '@/components/sections/how-it-works'
+import { Features } from '@/components/sections/features'
+import { Cli } from '@/components/sections/cli'
+import { Agents } from '@/components/sections/agents'
+import { Privacy } from '@/components/sections/privacy'
+import { Faq } from '@/components/sections/faq'
+import { Cta } from '@/components/sections/cta'
+import { LaunchBadgeTrack } from '@/components/sections/launch-badges'
+import { SiteFooter } from '@/components/sections/site-footer'
 
 export const Route = createFileRoute('/')({ component: LandingPage })
 
-const GITHUB_URL = 'https://github.com/limehq/munkel'
-const DOWNLOAD_URL = `${GITHUB_URL}/releases/latest`
-const PROTOCOL_URL = `${GITHUB_URL}/blob/main/apps/server/src/protocol.ts`
-const CLI_URL = `${GITHUB_URL}/tree/main/apps/cli`
-const LICENSE_URL = `${GITHUB_URL}/blob/main/LICENSE`
-const BREW_CMD = 'brew install limehq/tap/munkel'
-
-function GithubIcon({ className = '' }: { className?: string }) {
-  return (
-    <svg
-      className={`lucide ${className}`.trim()}
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      aria-hidden
-    >
-      <path d="M15 22v-4a4.8 4.8 0 0 0-1-3.5c3 0 6-2 6-5.5.08-1.25-.27-2.48-1-3.5.28-1.15.28-2.35 0-3.5 0 0-1 0-3 1.5-2.64-.5-5.36-.5-8 0C6 2 5 2 5 2c-.3 1.15-.3 2.35 0 3.5A5.403 5.403 0 0 0 4 9c0 3.5 3 5.5 6 5.5-.39.49-.68 1.05-.85 1.65-.17.6-.22 1.23-.15 1.85v4"></path>
-      <path d="M9 18c-4.51 2-5-2-7-2"></path>
-    </svg>
-  )
-}
-
-/* ---------- Announce bar: cross-platform / Windows soon ---------- */
-
-function AnnounceBar() {
-  return (
-    <a
-      className="announce"
-      href={GITHUB_URL}
-      aria-label="Munkel is going cross-platform — Windows support is coming soon"
-    >
-      <span className="announce-tag">Soon</span>
-      <span className="announce-text">
-        <MonitorSmartphone aria-hidden />
-        Going cross-platform — <strong>Windows support</strong> is on the way.
-      </span>
-      <ArrowRight className="announce-arrow" aria-hidden />
-    </a>
-  )
-}
-
-/* ---------- Nav: floating bar + sliding active pill ---------- */
-
-const NAV_LINKS = [
-  ['#how', 'How it works'],
-  ['#features', 'Features'],
-  ['#cli', 'CLI'],
-  ['#agents', 'Agents'],
-  ['#privacy', 'Privacy'],
-] as const
-
-type PillRect = { left: number; top: number; width: number; height: number }
-
-function Nav() {
-  const [floating, setFloating] = useState(false)
-  const [active, setActive] = useState<string | null>(null)
-  const [pill, setPill] = useState<PillRect | null>(null)
-  const linksRef = useRef<HTMLDivElement>(null)
-  const activeRef = useRef<string | null>(null)
-
-  // Hysteresis: enter floating well below the top, leave only near the
-  // very top — so the bar never flip-flops around a single threshold.
-  useEffect(() => {
-    const onScroll = () => {
-      const y = window.scrollY
-      if (y > 72) setFloating(true)
-      else if (y < 16) setFloating(false)
-    }
-    window.addEventListener('scroll', onScroll, { passive: true })
-    onScroll()
-    return () => window.removeEventListener('scroll', onScroll)
-  }, [])
-
-  const moveTo = (href: string) => {
-    const wrap = linksRef.current
-    if (!wrap) return
-    const a = wrap.querySelector<HTMLAnchorElement>(`a[href="${href}"]`)
-    if (!a) return
-    activeRef.current = href
-    setActive(href)
-    setPill({ left: a.offsetLeft, top: a.offsetTop, width: a.offsetWidth, height: a.offsetHeight })
-  }
-
-  useEffect(() => {
-    const onResize = () => {
-      if (activeRef.current) moveTo(activeRef.current)
-    }
-    window.addEventListener('resize', onResize)
-    if (location.hash && NAV_LINKS.some(([href]) => href === location.hash)) {
-      const hash = location.hash
-      requestAnimationFrame(() => moveTo(hash))
-    }
-    return () => window.removeEventListener('resize', onResize)
-  }, [])
-
-  const toggleTheme = () => {
-    const dark = document.documentElement.classList.toggle('dark')
-    try {
-      localStorage.setItem('munkel-theme', dark ? 'dark' : 'light')
-    } catch {
-      // private mode etc. — theme just won't persist
-    }
-  }
-
-  return (
-    <nav className={floating ? 'floating' : undefined}>
-      <div className="nav-inner">
-        <div className="nav-left">
-          <a href="/" className="wordmark">
-            <span className="dot"></span>munkel
-          </a>
-          <div className="nav-links" ref={linksRef}>
-            <span
-              className="nav-active-pill"
-              style={pill ? { ...pill, opacity: 1 } : undefined}
-            />
-            {NAV_LINKS.map(([href, label]) => (
-              <a
-                key={href}
-                href={href}
-                className={active === href ? 'active' : undefined}
-                onClick={() => moveTo(href)}
-              >
-                {label}
-              </a>
-            ))}
-          </div>
-        </div>
-        <div className="nav-actions">
-          <a className="nav-cta" href={DOWNLOAD_URL} aria-label="Download">
-            <Download aria-hidden />
-            <span>Download</span>
-          </a>
-          <a className="icon-btn" href={GITHUB_URL} aria-label="GitHub" title="GitHub">
-            <GithubIcon />
-          </a>
-          <button className="icon-btn" onClick={toggleTheme} aria-label="Toggle theme">
-            <Moon className="moon" aria-hidden />
-            <Sun className="sun" aria-hidden />
-          </button>
-        </div>
-      </div>
-    </nav>
-  )
-}
-
-/* ---------- Hero: pinned scroll stage + notch whisper demo ---------- */
-
-const MESSAGES = [
-  { name: 'Alex', avatar: '/avatars/01.png', text: 'coffee?' },
-  { name: 'Sam', avatar: '/avatars/02.png', text: 'on my way' },
-  { name: 'Taylor', avatar: '/avatars/03.png', text: 'deploy is live, go look' },
-  { name: 'Morgan', avatar: '/avatars/05.png', text: 'same table as last time' },
-]
-
-type NotchCtl = {
-  enabled: boolean
-  dwell: number
-  docked: boolean
-  /** The load-time whisper is showing — keeps the director's retract branch away at p ≈ 0. */
-  early: boolean
-  /** Pointer is over the notch — pauses rotation and defers the early-whisper close. */
-  hovered: boolean
-  openTeaser?: () => void
-  closeTeaser?: () => void
-}
-
-const sleep = (ms: number) => new Promise<void>((r) => setTimeout(r, ms))
-
-function Hero() {
-  const stageRef = useRef<HTMLDivElement>(null)
-  const copyRef = useRef<HTMLDivElement>(null)
-  const wrapRef = useRef<HTMLDivElement>(null)
-  const macRef = useRef<HTMLDivElement>(null)
-  const screenRef = useRef<HTMLDivElement>(null)
-  const hintRef = useRef<HTMLDivElement>(null)
-  const demoHintRef = useRef<HTMLDivElement>(null)
-  const notchRef = useRef<HTMLDivElement>(null)
-  const msgRef = useRef<HTMLDivElement>(null)
-  const avatarRef = useRef<HTMLImageElement>(null)
-  const notchCtl = useRef<NotchCtl>({
-    enabled: true,
-    dwell: 3400,
-    docked: false,
-    early: false,
-    hovered: false,
-  })
-  const [clock, setClock] = useState('Thu 9:41')
-  const [msgCopied, setMsgCopied] = useState(false)
-
-  useEffect(() => {
-    const now = new Date()
-    setClock(
-      now.toLocaleDateString('en-US', { weekday: 'short' }) +
-        ' ' +
-        now.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit', hour12: false }),
-    )
-  }, [])
-
-  // Notch demo: opened by the scroll director the moment the MacBook
-  // docks. Once open, it stays open; messages rotate in place until
-  // closeTeaser(). The session token guarantees only ONE rotation loop
-  // is ever alive — a stale loop sleeping through a close/reopen cycle
-  // sees a newer session and exits.
-  useEffect(() => {
-    const root = notchRef.current
-    const elMsg = msgRef.current
-    const elAvatar = avatarRef.current
-    if (!root || !elMsg || !elAvatar) return
-
-    let open = false
-    let session = 0
-    let i = 0
-
-    // Hover only exists on fine pointers — on touch, a tap emulates
-    // mouseenter with no matching mouseleave and would freeze the
-    // rotation forever. Touch gets a longer dwell instead.
-    const canHover = window.matchMedia('(hover: hover)').matches
-    if (!canHover) notchCtl.current.dwell = 5000
-    const onEnter = () => {
-      notchCtl.current.hovered = true
-      // Dismiss the hint only once it has actually been seen — a hover on
-      // the early whisper must not eat the caption before dock ever shows it.
-      const cap = demoHintRef.current
-      if (cap?.classList.contains('show')) cap.classList.add('off')
-    }
-    const onLeave = () => {
-      notchCtl.current.hovered = false
-    }
-    if (canHover) {
-      root.addEventListener('mouseenter', onEnter)
-      root.addEventListener('mouseleave', onLeave)
-    }
-
-    const ctl = notchCtl.current
-    ctl.openTeaser = async () => {
-      if (open) return
-      open = true
-      const s = ++session
-      const live = () => open && s === session
-      const m = MESSAGES[i % MESSAGES.length]
-      i++
-      elMsg.textContent = m.text
-      elAvatar.src = m.avatar
-      // A reopen inside retract()'s 260ms fallback window must not inherit
-      // its fast inline transitions — reset to the CSS spring first.
-      root.style.transition = ''
-      root.querySelectorAll<HTMLElement>('.mbn-avatar, .mbn-copy, .mbn-body').forEach((e) => {
-        e.style.transition = ''
-      })
-      root.classList.add('teaser')
-      // Robustness: if the CSS transition is throttled/frozen (background
-      // tab, energy saver), snap to the final teaser state inline.
-      await sleep(600)
-      if (!live()) return
-      if (parseFloat(getComputedStyle(root).width) < 200) {
-        const small = window.matchMedia('(max-width: 900px)').matches
-        root.style.transition = 'none'
-        root.style.width = small ? '220px' : '260px'
-        root.style.height = small ? '44px' : '56px'
-        root.style.borderRadius = '0 0 18px 18px'
-        root.querySelectorAll<HTMLElement>('.mbn-avatar, .mbn-copy, .mbn-body').forEach((e) => {
-          e.style.transition = 'none'
-          e.style.opacity = '1'
-        })
-      }
-      while (live()) {
-        await sleep(ctl.dwell)
-        while (ctl.hovered && live()) {
-          await sleep(300)
-        }
-        if (!live()) break
-        const n = MESSAGES[i % MESSAGES.length]
-        i++
-        elMsg.style.transition = 'opacity 0.18s ease'
-        elAvatar.style.transition = 'opacity 0.18s ease'
-        elMsg.style.opacity = '0'
-        elAvatar.style.opacity = '0'
-        await sleep(200)
-        if (!live()) return
-        elMsg.textContent = n.text
-        elAvatar.src = n.avatar
-        elMsg.style.opacity = '1'
-        elAvatar.style.opacity = '1'
-        await sleep(220)
-        if (!live()) return
-        elMsg.style.transition = ''
-        elAvatar.style.transition = ''
-        elMsg.style.opacity = ''
-        elAvatar.style.opacity = ''
-      }
-    }
-    // Called by the scroll director when the user scrolls back above the
-    // dock point (the director handles the visual retract itself).
-    ctl.closeTeaser = () => {
-      open = false
-      session++
-      elMsg.style.transition = ''
-      elAvatar.style.transition = ''
-      elMsg.style.opacity = ''
-      elAvatar.style.opacity = ''
-    }
-
-    return () => {
-      open = false
-      session++
-      ctl.openTeaser = undefined
-      ctl.closeTeaser = undefined
-      root.removeEventListener('mouseenter', onEnter)
-      root.removeEventListener('mouseleave', onLeave)
-    }
-  }, [])
-
-  // Scroll director — one scrubbed timeline: copy recedes → MacBook takes
-  // center stage → zoom toward the notch → demo starts.
-  useEffect(() => {
-    const stage = stageRef.current
-    const copy = copyRef.current
-    const wrap = wrapRef.current
-    const mac = macRef.current
-    const screen = screenRef.current
-    const hint = hintRef.current
-    const mbn = notchRef.current
-    if (!stage || !copy || !wrap || !mac || !screen || !mbn) return
-    if (window.matchMedia('(prefers-reduced-motion: reduce)').matches) {
-      // No scrubbed timeline — but the demo must still exist: show the
-      // first whisper statically (the JSX already carries the message).
-      mbn.classList.add('teaser')
-      return
-    }
-
-    const ctl = notchCtl.current
-    const clamp01 = (t: number) => Math.max(0, Math.min(1, t))
-    const ease = (t: number) => (t < 0.5 ? 2 * t * t : 1 - Math.pow(-2 * t + 2, 2) / 2)
-
-    // Fast visual retract — shared by the scroll-back branch and the
-    // early-whisper timer below.
-    const retract = () => {
-      if (ctl.closeTeaser) ctl.closeTeaser()
-      mbn!.style.transition = 'width 0.18s ease, height 0.18s ease, border-radius 0.18s ease'
-      mbn!.classList.remove('teaser')
-      mbn!.style.width = ''
-      mbn!.style.height = ''
-      mbn!.style.borderRadius = ''
-      mbn!.querySelectorAll<HTMLElement>('.mbn-avatar, .mbn-copy, .mbn-body').forEach((el) => {
-        el.style.transition = 'opacity 0.1s ease, visibility 0s linear 0.1s'
-        el.style.opacity = ''
-      })
-      setTimeout(() => {
-        if (mbn!.classList.contains('teaser')) return
-        // Frozen-timeline fallback: if the retract transition never ran,
-        // snap to the closed state with transitions off.
-        if (parseFloat(getComputedStyle(mbn!).width) > 200) {
-          mbn!.style.transition = 'none'
-          mbn!
-            .querySelectorAll<HTMLElement>('.mbn-avatar, .mbn-copy, .mbn-body')
-            .forEach((el) => {
-              el.style.transition = 'none'
-            })
-          void mbn!.offsetWidth
-          requestAnimationFrame(() => {
-            mbn!.style.transition = ''
-            mbn!
-              .querySelectorAll<HTMLElement>('.mbn-avatar, .mbn-copy, .mbn-body')
-              .forEach((el) => {
-                el.style.transition = ''
-              })
-          })
-        } else {
-          mbn!.style.transition = ''
-          mbn!
-            .querySelectorAll<HTMLElement>('.mbn-avatar, .mbn-copy, .mbn-body')
-            .forEach((el) => {
-              el.style.transition = ''
-            })
-        }
-      }, 260)
-    }
-
-    let ticking = false
-    function update() {
-      ticking = false
-      const vh = window.innerHeight
-      const total = Math.max(1, stage!.offsetHeight - vh)
-      const p = clamp01(-stage!.getBoundingClientRect().top / total)
-
-      // 1 — hero copy recedes and fades out
-      const pc = ease(clamp01(p / 0.38))
-      copy!.style.opacity = (1 - pc).toFixed(3)
-      copy!.style.transform =
-        'translateY(' + (-64 * pc).toFixed(1) + 'px) scale(' + (1 - 0.05 * pc).toFixed(4) + ')'
-      copy!.style.pointerEvents = pc > 0.5 ? 'none' : ''
-      if (hint) hint.style.opacity = Math.max(0, 1 - p / 0.06).toFixed(3)
-
-      // 2 — the MacBook takes the stage: rises to center, flattens, grows
-      const pm = ease(clamp01((p - 0.04) / 0.6))
-      const targetTop = (vh - wrap!.offsetHeight) / 2
-      const ty = (targetTop - wrap!.offsetTop) * pm
-      const sc = 0.94 + 0.2 * pm
-      wrap!.style.transform = 'translateY(' + ty.toFixed(1) + 'px) scale(' + sc.toFixed(4) + ')'
-      // 3 — keep scrolling: zoom in, docked to the MacBook's top edge
-      // (transform-origin 50% 0%, so the notch edge stays pinned)
-      const pz = ease(clamp01((p - 0.62) / 0.38))
-      mac!.style.transform =
-        'rotateX(' + (22 * (1 - pm)).toFixed(2) + 'deg) scale(' + (1 + 0.45 * pz).toFixed(4) + ')'
-      mac!.style.opacity = (0.55 + 0.45 * pm).toFixed(3)
-      // The 1px rim hairline aliases in/out under non-integer scale —
-      // kill it completely BEFORE the zoom starts so it can never flicker.
-      const pr = ease(clamp01((p - 0.48) / 0.14))
-      screen!.style.borderColor = 'oklch(1 0 0 / ' + (0.16 * (1 - pr)).toFixed(3) + ')'
-
-      // 4 — once the MacBook holds the stage, the demo whispers — opened
-      // directly from here so a fast scroll can never miss it.
-      ctl.docked = pm >= 0.95
-      if (ctl.docked) {
-        // Docking takes over from the early whisper.
-        ctl.early = false
-        mac!.classList.remove('early')
-        if (ctl.enabled && ctl.openTeaser) ctl.openTeaser()
-        demoHintRef.current?.classList.add('show')
-      } else {
-        demoHintRef.current?.classList.remove('show')
-        // Scrolling back mid-message: retract the teaser fast.
-        if (!ctl.early && mbn!.classList.contains('teaser')) retract()
-      }
-    }
-    function onScroll() {
-      if (!ticking) {
-        ticking = true
-        requestAnimationFrame(update)
-      }
-    }
-    window.addEventListener('scroll', onScroll, { passive: true })
-    window.addEventListener('resize', onScroll, { passive: true })
-    update()
-
-    // A literal "Psst": one whisper above the fold shortly after load, so
-    // the demo exists before anyone scrolls. ctl.early keeps the director's
-    // retract branch from snapping it shut at p ≈ 0; .early lifts the
-    // dimmed, tilted MacBook so the message is legible.
-    const tOpen = window.setTimeout(() => {
-      if (ctl.docked) return
-      ctl.early = true
-      mac!.classList.add('early')
-      if (ctl.enabled && ctl.openTeaser) ctl.openTeaser()
-    }, 1400)
-    let tClose = 0
-    const closeEarly = () => {
-      if (!ctl.early) {
-        mac!.classList.remove('early')
-        return
-      }
-      // Hovering holds the whisper open — just like the real notch.
-      if (ctl.hovered) {
-        tClose = window.setTimeout(closeEarly, 900)
-        return
-      }
-      mac!.classList.remove('early')
-      ctl.early = false
-      if (!ctl.docked) retract()
-    }
-    tClose = window.setTimeout(closeEarly, 4600)
-
-    return () => {
-      window.removeEventListener('scroll', onScroll)
-      window.removeEventListener('resize', onScroll)
-      clearTimeout(tOpen)
-      clearTimeout(tClose)
-    }
-  }, [])
-
-  const copyMessage = () => {
-    const text = msgRef.current?.textContent ?? ''
-    if (navigator.clipboard) navigator.clipboard.writeText(text).catch(() => {})
-    setMsgCopied(true)
-    setTimeout(() => setMsgCopied(false), 1400)
-  }
-
-  return (
-    <header className="hero">
-      <div className="hero-stage" ref={stageRef}>
-        <div className="hero-sticky">
-          <div className="container hero-copy" ref={copyRef}>
-            <div className="hero-kicker">Ephemeral circle messages for macOS</div>
-            <h1>
-              Psst, your <span className="notch-word">notch</span> has something to tell&nbsp;you.
-            </h1>
-            <p className="lead">
-              Quick pings between friends, end-to-end encrypted — they slide out of the notch,
-              linger a moment, and vanish. Nothing is ever stored.
-            </p>
-            <div className="hero-ctas">
-              <a className="btn btn-primary" href={DOWNLOAD_URL}>
-                <Download aria-hidden />
-                Download for macOS
-              </a>
-              <a className="btn btn-outline" href={GITHUB_URL}>
-                <GithubIcon />
-                View on GitHub
-              </a>
-            </div>
-            <div className="hero-meta">
-              Free &amp; open source · macOS 14+ · Apple Silicon &amp; Intel · Signed &amp;
-              notarized · works without a notch, too
-            </div>
-          </div>
-          <div className="mockup-wrap" ref={wrapRef}>
-            <div className="macbook" ref={macRef}>
-              <div className="mb-screen" ref={screenRef}>
-                <div className="mb-display">
-                  <div className="mb-wallpaper"></div>
-                  <div className="mb-menubar">
-                    <div className="mb-menu">
-                      <span className="mb-app">munkel</span>
-                      <span>Circles</span>
-                      <span>Identity</span>
-                      <span>Help</span>
-                    </div>
-                    <div className="mb-menu-right">
-                      <Wifi aria-hidden />
-                      <span>{clock}</span>
-                    </div>
-                  </div>
-                  <div
-                    className="mb-notch"
-                    ref={notchRef}
-                    onClick={() => {
-                      // In the page demo a click copies the message (the real
-                      // notch opens an inline reply on click).
-                      if (notchRef.current?.classList.contains('teaser')) copyMessage()
-                    }}
-                  >
-                    <span className="notch-cam"></span>
-                    <img className="mbn-avatar" ref={avatarRef} src="/avatars/01.png" alt="" />
-                    <button
-                      className={`mbn-copy${msgCopied ? ' copied' : ''}`}
-                      onClick={copyMessage}
-                      aria-label="Copy message"
-                    >
-                      <Copy className="ic-copy" strokeWidth={2} aria-hidden />
-                      <Check className="ic-check" strokeWidth={2.5} aria-hidden />
-                    </button>
-                    <div className="mbn-body">
-                      <div className="mbn-msg" ref={msgRef}>
-                        coffee?
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="demo-caption" ref={demoHintRef} aria-hidden>
-            hover to hold a munkel open — just like the real thing
-          </div>
-          <div className="scroll-hint" ref={hintRef}>
-            <span>See it munkel</span>
-            <ChevronDown aria-hidden />
-          </div>
-        </div>
-      </div>
-    </header>
-  )
-}
-
-/* ---------- Terminal typing demo ---------- */
-
-type TermStep = { type: 'cmd'; text: string } | { type: 'out'; html: string }
-
-// Mirrors the real CLI output (apps/cli/src/munkel.ts): circles print as
-// `● code  —  members`, every successful send prints `munkeled ✓`.
-const TERM_SCRIPT: TermStep[] = [
-  { type: 'cmd', text: 'munkel circles' },
-  { type: 'out', html: '<span class="tdot-live">●</span> blue-table-42  <span class="tdim">—</span>  Alex, Sam, Morgan' },
-  { type: 'out', html: '<span class="tdot-live">●</span> project-7  <span class="tdim">—</span>  Sam, Alex' },
-  { type: 'cmd', text: 'munkel blue-table-42 all "table\'s free, come down"' },
-  { type: 'out', html: '<span class="tdim">munkeled ✓</span>' },
-  { type: 'cmd', text: 'munkel project-7 Sam "package for you downstairs"' },
-  { type: 'out', html: '<span class="tdim">munkeled ✓</span>' },
-]
-
-function TerminalDemo({ onFirstSend }: { onFirstSend?: () => void }) {
-  const termRef = useRef<HTMLDivElement>(null)
-  const bodyRef = useRef<HTMLDivElement>(null)
-  const onFirstSendRef = useRef(onFirstSend)
-  onFirstSendRef.current = onFirstSend
-
-  useEffect(() => {
-    const term = termRef.current
-    const body = bodyRef.current
-    if (!term || !body) return
-
-    let started = false
-    let cancelled = false
-    let sentFired = false
-    // Under reduced motion the script renders instantly instead of typing.
-    const instant = window.matchMedia('(prefers-reduced-motion: reduce)').matches
-
-    async function run() {
-      body!.innerHTML = ''
-      for (const step of TERM_SCRIPT) {
-        if (cancelled) return
-        const line = document.createElement('div')
-        line.className = 'tline'
-        body!.appendChild(line)
-        if (step.type === 'cmd') {
-          line.innerHTML =
-            '<span class="tprompt">$ </span><span class="ttext"></span><span class="cursor"></span>'
-          const ttext = line.querySelector('.ttext')!
-          if (instant) {
-            ttext.textContent = step.text
-          } else {
-            for (const ch of step.text) {
-              if (cancelled) return
-              ttext.textContent += ch
-              await sleep(28 + Math.random() * 40)
-            }
-            await sleep(350)
-          }
-          if (cancelled) return
-          line.querySelector('.cursor')?.remove()
-        } else {
-          line.innerHTML = step.html
-          if (!sentFired && step.html.includes('munkeled')) {
-            sentFired = true
-            onFirstSendRef.current?.()
-          }
-          if (!instant) await sleep(420)
-        }
-      }
-      if (cancelled) return
-      const done = document.createElement('div')
-      done.className = 'tline'
-      done.innerHTML = '<span class="tprompt">$ </span><span class="cursor"></span>'
-      body!.appendChild(done)
-    }
-
-    function start() {
-      if (!started) {
-        started = true
-        run()
-      }
-    }
-    function checkVisible() {
-      const rect = term!.getBoundingClientRect()
-      if (rect.top < window.innerHeight * 0.85 && rect.bottom > 0) start()
-    }
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting) start()
-      },
-      { threshold: 0.4 },
-    )
-    observer.observe(term)
-    window.addEventListener('scroll', checkVisible, { passive: true })
-    checkVisible()
-    const fallback = setTimeout(start, 12000)
-
-    return () => {
-      cancelled = true
-      observer.disconnect()
-      window.removeEventListener('scroll', checkVisible)
-      clearTimeout(fallback)
-    }
-  }, [])
-
-  return (
-    <div className="terminal" ref={termRef}>
-      <div className="terminal-bar">
-        <span className="tdot"></span>
-        <span className="tdot"></span>
-        <span className="tdot"></span>
-        <span className="ttitle">zsh · munkel</span>
-      </div>
-      <div className="terminal-body" ref={bodyRef}>
-        <span className="cursor"></span>
-      </div>
-    </div>
-  )
-}
-
-/* ---------- CLI demo + the receiving notch ---------- */
-
-function CliShowcase() {
-  const [sent, setSent] = useState(false)
-  return (
-    <div className="cli-stage">
-      <TerminalDemo onFirstSend={() => setSent(true)} />
-      <div className={`meanwhile${sent ? ' show' : ''}`} aria-hidden={!sent}>
-        <div className="meanwhile-caption">meanwhile, on Alex's Mac</div>
-        <div className="mini-notch">
-          <img src="/avatars/03.png" alt="" />
-          <div style={{ display: 'flex', flexDirection: 'column' }}>
-            <span className="mn-name">Taylor</span>
-            <span>table's free, come down</span>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-/* ---------- Install command switcher (Agents) ---------- */
-
-const INSTALL_CMDS = {
-  npx: 'npx skills add limehq/munkel',
-  pnpm: 'pnpm dlx skills add limehq/munkel',
-  yarn: 'yarn dlx skills add limehq/munkel',
-  bun: 'bunx skills add limehq/munkel',
-} as const
-
-type Pm = keyof typeof INSTALL_CMDS
-
-function InstallCmd() {
-  const [pm, setPm] = useState<Pm>('npx')
-  const [copied, setCopied] = useState(false)
-
-  const copy = () => {
-    if (navigator.clipboard) navigator.clipboard.writeText(INSTALL_CMDS[pm]).catch(() => {})
-    setCopied(true)
-    setTimeout(() => setCopied(false), 1400)
-  }
-
-  return (
-    <div className="install-cmd">
-      <div className="install-bar">
-        <div className="install-tabs" role="tablist">
-          {(Object.keys(INSTALL_CMDS) as Pm[]).map((key) => (
-            <button
-              key={key}
-              className={pm === key ? 'active' : undefined}
-              role="tab"
-              aria-selected={pm === key}
-              onClick={() => setPm(key)}
-            >
-              {key}
-            </button>
-          ))}
-        </div>
-        <button
-          className={`install-copy${copied ? ' copied' : ''}`}
-          onClick={copy}
-          aria-label="Copy command"
-        >
-          <Copy className="ic-copy" strokeWidth={2} aria-hidden />
-          <Check className="ic-check" strokeWidth={2.5} aria-hidden />
-        </button>
-      </div>
-      <div className="install-body">
-        <span className="prompt">$</span>
-        <span>{INSTALL_CMDS[pm]}</span>
-      </div>
-    </div>
-  )
-}
-
-/* ---------- Brew one-liner (final CTA) ---------- */
-
-function BrewCmd() {
-  const [copied, setCopied] = useState(false)
-
-  const copy = () => {
-    if (navigator.clipboard) navigator.clipboard.writeText(BREW_CMD).catch(() => {})
-    setCopied(true)
-    setTimeout(() => setCopied(false), 1400)
-  }
-
-  return (
-    <div className="install-cmd brew-cmd">
-      <div className="install-body">
-        <span className="prompt">$</span>
-        <span>{BREW_CMD}</span>
-        <button
-          className={`install-copy${copied ? ' copied' : ''}`}
-          onClick={copy}
-          aria-label="Copy install command"
-        >
-          <Copy className="ic-copy" strokeWidth={2} aria-hidden />
-          <Check className="ic-check" strokeWidth={2.5} aria-hidden />
-        </button>
-      </div>
-    </div>
-  )
-}
-
-/* ---------- Launch badge track ---------- */
-
-// Platforms Munkel launches on. Until we actually land on one, its badge is
-// rendered in-house (the "Launching on" pill). When the listing goes live, set
-// `live: true`, point `href` at our listing, and drop the platform's official
-// badge into `img` — self-hosted under /badges so we never hot-link a third
-// party (no layout shift, no tracking, no availability risk). See issue #8.
-type LaunchPlatform = {
-  name: string
-  href: string
-  live?: boolean
-  /** Self-hosted official badge, e.g. '/badges/product-hunt.svg'. */
-  img?: string
-  alt?: string
-}
-
-// Source of truth + status tracking: docs/launch-platforms.md. This array
-// mirrors the "live" rows from there; hrefs are platform homepages until we
-// have our own listing URL for each.
-const LAUNCH_PLATFORMS: LaunchPlatform[] = [
-  { name: 'Product Hunt', href: 'https://www.producthunt.com' },
-  { name: 'Peerlist', href: 'https://peerlist.io' },
-  { name: 'Fazier', href: 'https://fazier.com' },
-  { name: 'Uneed', href: 'https://www.uneed.best' },
-  { name: 'MicroLaunch', href: 'https://microlaunch.net' },
-  { name: 'Startup Fame', href: 'https://startupfa.me' },
-  { name: 'Twelve Tools', href: 'https://twelve.tools' },
-  { name: 'Findly.tools', href: 'https://findly.tools' },
-  { name: 'LaunchIgniter', href: 'https://launchigniter.com' },
-  { name: 'Dang.ai', href: 'https://dang.ai' },
-]
-
-function LaunchBadge({ platform }: { platform: LaunchPlatform }) {
-  const { name, href, live, img, alt } = platform
-  return (
-    <a
-      className="launch-badge"
-      href={href}
-      target="_blank"
-      rel="noopener"
-      aria-label={live ? `Featured on ${name}` : `Munkel is launching on ${name}`}
-    >
-      {img ? (
-        <img src={img} alt={alt ?? `Featured on ${name}`} loading="lazy" decoding="async" />
-      ) : (
-        <>
-          <span className="lb-glyph">
-            <Rocket aria-hidden />
-          </span>
-          <span className="lb-text">
-            <span className="lb-kicker">{live ? 'Featured on' : 'Launching on'}</span>
-            <span className="lb-name">{name}</span>
-          </span>
-        </>
-      )}
-    </a>
-  )
-}
-
-// A seamless marquee: the badge list is rendered twice (the duplicate is
-// display:contents, so its badges join the same flex row with even spacing) and
-// the row is translated by exactly -50%, looping without a seam. Under reduced
-// motion the duplicate is hidden and the badges wrap statically. Pass liveOnly
-// once we have real launches to shrink the strip to earned badges.
-function LaunchBadgeTrack({ liveOnly = false }: { liveOnly?: boolean }) {
-  const platforms = liveOnly ? LAUNCH_PLATFORMS.filter((p) => p.live) : LAUNCH_PLATFORMS
-  if (platforms.length === 0) return null
-  return (
-    <section className="badge-track">
-      <div className="container">
-        <div className="section-kicker">Launch</div>
-        <h2>Out in the wild.</h2>
-        <p>
-          Munkel is rolling out across the platforms where people hunt for new tools. Catch the
-          launch — and give us an upvote when we land.
-        </p>
-      </div>
-      <div className="badge-marquee">
-        <div className="badge-marquee-row">
-          {platforms.map((p) => (
-            <LaunchBadge key={p.name} platform={p} />
-          ))}
-          <span className="badge-dup" aria-hidden>
-            {platforms.map((p) => (
-              <LaunchBadge key={`${p.name}-dup`} platform={p} />
-            ))}
-          </span>
-        </div>
-      </div>
-    </section>
-  )
-}
-
-/* ---------- Page ---------- */
-
 function LandingPage() {
   return (
-    <>
+    <MotionConfig reducedMotion="user">
       <AnnounceBar />
       <Nav />
       <Hero />
-
-      <section id="how">
-        <div className="container">
-          <div className="section-head">
-            <div className="section-kicker">How it works</div>
-            <h2>A circle is born from a code.</h2>
-            <p>
-              Group chats want your attention — badges to clear, threads to catch up on. Munkel
-              doesn't: a message glances past in the notch and is gone.
-            </p>
-            <p>
-              No invites, no server-side circle state. Sign in with GitHub once, then three steps and
-              you're Munkeling.
-            </p>
-          </div>
-          <div className="steps">
-            <div className="step">
-              <span className="step-num">01</span>
-              <h3>Create a circle</h3>
-              <p>
-                The app mints a human-readable code. Say it across the table or paste it in a chat.
-                That's the whole onboarding.
-              </p>
-              <div className="step-visual">
-                <span style={{ fontFamily: 'var(--font-mono)', fontSize: 14, color: 'var(--brand)' }}>
-                  blue-table-42
-                </span>
-              </div>
-            </div>
-            <div className="step">
-              <span className="step-num">02</span>
-              <h3>Friends join</h3>
-              <p>
-                Anyone with the code is in. The code doubles as the AES-256-GCM key, so the relay
-                only ever routes opaque blobs.
-              </p>
-              <div className="step-visual">
-                <div className="avatar-stack">
-                  <img src="/avatars/01.png" alt="Alex" />
-                  <img src="/avatars/02.png" alt="Sam" />
-                  <img src="/avatars/05.png" alt="Morgan" />
-                  <span className="joined">3 joined</span>
-                </div>
-              </div>
-            </div>
-            <div className="step">
-              <span className="step-num">03</span>
-              <h3>Read the notch</h3>
-              <p>
-                Messages slide out, linger, disappear. Hover to hold one open, click to reply
-                right in the notch — it closes again after you send.
-              </p>
-              <div className="step-visual">
-                <div className="mini-notch">
-                  <img src="/avatars/01.png" alt="" />
-                  <div style={{ display: 'flex', flexDirection: 'column' }}>
-                    <span className="mn-name">Alex</span>
-                    <span>down in 5</span>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <section id="features">
-        <div className="container">
-          <div className="section-head">
-            <div className="section-kicker">Features</div>
-            <h2>Small surface. Sharp edges.</h2>
-            <p>Everything Munkel does, and the things it refuses to do.</p>
-          </div>
-          <div className="features">
-            <div className="feature">
-              <div className="feature-icon">
-                <TimerOff aria-hidden />
-              </div>
-              <h3>Ephemeral by design</h3>
-              <p>
-                The relay holds zero state: no database, no logs of content. Messages exist only in
-                flight; offline means missed, like a real munkel.
-              </p>
-            </div>
-            <div className="feature">
-              <div className="feature-icon">
-                <Lock aria-hidden />
-              </div>
-              <h3>End-to-end encrypted</h3>
-              <p>
-                AES-256-GCM, key derived from the circle code on-device. The relay routes ciphertext
-                it cannot open. By construction, not by promise.
-              </p>
-            </div>
-            <div className="feature">
-              <div className="feature-icon">
-                <UserRoundX aria-hidden />
-              </div>
-              <h3>No accounts of our own</h3>
-              <p>
-                No email, no phone number, no password. You sign in once with GitHub; Munkel runs no
-                identity service and keeps nothing server-side.
-              </p>
-            </div>
-            <div className="feature">
-              <div className="feature-icon">
-                <GithubIcon />
-              </div>
-              <h3>GitHub identity</h3>
-              <p>
-                Your display name and avatar come from GitHub via device flow, one login at setup.
-                The token is requested with empty scope, used once, then discarded.
-              </p>
-            </div>
-            <div className="feature">
-              <div className="feature-icon">
-                <Terminal aria-hidden />
-              </div>
-              <h3>munkel CLI</h3>
-              <p>
-                Send from your shell, script it, or wire it into agents over the app's control
-                socket. An MCP server is on the way.
-              </p>
-            </div>
-            <div className="feature">
-              <div className="feature-icon">
-                <Laptop aria-hidden />
-              </div>
-              <h3>Works without a notch</h3>
-              <p>
-                Older MacBook or external display? Messages fall back to an elegant floating panel.
-                Same munkel, different hardware.
-              </p>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <section id="cli">
-        <div className="container">
-          <div className="cli-grid">
-            <div className="cli-copy">
-              <div className="section-kicker">CLI</div>
-              <h2>Munkel from the shell.</h2>
-              <p>
-                <span className="code">munkel</span> is a thin client over the app's Unix domain
-                socket. The app owns all crypto and relay connections, the CLI just talks.
-                Recipients by display name, circles by code prefix.
-              </p>
-              <p>
-                Newline-delimited JSON over <span className="code">control.sock</span> makes it an
-                ideal substrate for scripts and agents.
-              </p>
-            </div>
-            <CliShowcase />
-          </div>
-        </div>
-      </section>
-
-      <section id="agents">
-        <div className="container">
-          <div className="section-head">
-            <div className="section-kicker">Agents</div>
-            <h2>LLM ready.</h2>
-            <p>
-              The CLI is plain text in, plain text out. Anything that can run a shell can munkel,
-              including the agent you already use.
-            </p>
-          </div>
-          <InstallCmd />
-          <div className="features cols-2">
-            <div className="feature">
-              <div className="feature-icon">
-                <Bot aria-hidden />
-              </div>
-              <h3>Agents can send for real</h3>
-              <p>
-                An LLM with shell access uses <span className="code">munkel</span> exactly like you
-                do: pick a person or a circle, send the message. "Tell blue-table-42 I'm running
-                late" becomes one command.
-              </p>
-            </div>
-            <div className="feature">
-              <div className="feature-icon">
-                <Package aria-hidden />
-              </div>
-              <h3>Skills, ready to install</h3>
-              <p>
-                Prepared skills teach your agent munkel's commands in one step. Install once and
-                your assistant knows how to munkel.
-              </p>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <section id="privacy">
-        <div className="container">
-          <div className="section-head">
-            <div className="section-kicker">Privacy</div>
-            <h2>What the relay sees.</h2>
-            <p>
-              One Durable Object per circle, WebSocket hibernation, no storage. Ephemerality is
-              enforced by architecture, not policy. The relay knows nothing.
-            </p>
-          </div>
-          <div className="privacy-banner">
-            <div className="privacy-banner-head">
-              <div className="feature-icon">
-                <MonitorOff aria-hidden />
-              </div>
-              <div>
-                <h3>Invisible to screen sharing</h3>
-                <p>
-                  Presenting in Zoom, Teams, or Meet? Munkel excludes its windows from screen
-                  capture. Munkels stay on your screen, never on the shared one.
-                </p>
-              </div>
-            </div>
-            <div className="share-compare" aria-hidden>
-              <figure className="share-panel">
-                <div className="share-screen">
-                  <div className="share-menubar"></div>
-                  <div className="share-notch open">
-                    <img src="/avatars/01.png" alt="" />
-                    <span>down in 5</span>
-                  </div>
-                </div>
-                <figcaption>Your screen</figcaption>
-              </figure>
-              <figure className="share-panel">
-                <div className="share-screen">
-                  <div className="share-menubar"></div>
-                  <div className="share-notch"></div>
-                </div>
-                <figcaption>What Zoom sees</figcaption>
-              </figure>
-            </div>
-          </div>
-          <div className="privacy-grid">
-            <div className="privacy-card sees">
-              <h3>The relay sees</h3>
-              <ul>
-                <li>
-                  <Info aria-hidden />
-                  <span className="li-text">
-                    Opaque encrypted blobs <span className="li-sub">· ciphertext it has no key for</span>
-                  </span>
-                </li>
-                <li>
-                  <Info aria-hidden />
-                  <span className="li-text">
-                    A group ID derived from your code <span className="li-sub">· not the code itself</span>
-                  </span>
-                </li>
-                <li>
-                  <Info aria-hidden />
-                  <span className="li-text">
-                    Connection timing <span className="li-sub">· that someone is online, not who</span>
-                  </span>
-                </li>
-              </ul>
-            </div>
-            <div className="privacy-card never">
-              <h3>The relay never sees</h3>
-              <ul>
-                <li>
-                  <Check strokeWidth={2} aria-hidden />
-                  <span className="li-text">
-                    Message contents <span className="li-sub">· encrypted end-to-end</span>
-                  </span>
-                </li>
-                <li>
-                  <Check strokeWidth={2} aria-hidden />
-                  <span className="li-text">
-                    Names &amp; avatars <span className="li-sub">· profiles travel inside encrypted payloads</span>
-                  </span>
-                </li>
-                <li>
-                  <Check strokeWidth={2} aria-hidden />
-                  <span className="li-text">
-                    Any history <span className="li-sub">· nothing is ever written to storage</span>
-                  </span>
-                </li>
-              </ul>
-            </div>
-          </div>
-          <div className="honest">
-            <p>
-              Munkel phones home to no one. The app opens exactly two kinds of connections: the
-              relay — which sees only what's listed above — and GitHub, once, at login. No
-              telemetry, no analytics, no crash reporting.
-            </p>
-            <p>
-              On your Mac, circle codes and your profile live in the app's local settings — a code
-              is never transmitted, to the relay or anyone else. The GitHub token exists only in
-              memory for one profile fetch and is never written to disk.
-            </p>
-            <p>
-              Honest limits: GitHub sees the device-flow login happen and lists Munkel under your
-              authorized apps. Profiles are display-only; peers get no cryptographic proof that a
-              member owns the GitHub name they show.
-            </p>
-            <p>
-              The wire protocol is specified where it is enforced:{' '}
-              <a href={PROTOCOL_URL}>protocol.ts</a>.
-            </p>
-          </div>
-        </div>
-      </section>
-
-      <section id="faq">
-        <div className="container">
-          <div className="section-head">
-            <div className="section-kicker">FAQ</div>
-            <h2>Quick answers.</h2>
-          </div>
-          <div className="faq">
-            <details>
-              <summary>Apple Silicon or Intel?</summary>
-              <p>Both. The app and the CLI ship as one universal binary — arm64 and x86_64.</p>
-            </details>
-            <details>
-              <summary>Signed and notarized?</summary>
-              <p>
-                Yes. Every release is Developer ID-signed and notarized by Apple, so Gatekeeper
-                opens it without warnings.
-              </p>
-            </details>
-            <details>
-              <summary>Electron or native?</summary>
-              <p>Native Swift and SwiftUI. One small menu-bar binary, no embedded browser.</p>
-            </details>
-            <details>
-              <summary>Can I reply from the notch?</summary>
-              <p>
-                Yes — click a munkel and reply right there; the notch closes again after you
-                send. You can also reply from the menu-bar popover or the{' '}
-                <span className="code">munkel</span> CLI.
-              </p>
-            </details>
-            <details>
-              <summary>How do I update?</summary>
-              <p>
-                Installed with Homebrew: <span className="code">brew upgrade munkel</span>.
-                Grabbed the zip: download the latest release and swap the app. There is no
-                auto-updater.
-              </p>
-            </details>
-          </div>
-        </div>
-      </section>
-
-      <section className="cta">
-        <div className="container">
-          <div className="founder-note">
-            <p>
-              "Munkel exists because group chats kept turning into obligations. It is meant for the
-              kind of note you would say across a table: brief, visible, and easy to let go."
-            </p>
-            <span className="founder-sig">Built for lightweight, low-pressure messages.</span>
-          </div>
-          <h2>Start Munkeling.</h2>
-          <p>
-            Open source, MIT licensed. A native Swift app — no Electron — one signed &amp;
-            notarized binary in your menu bar.
-          </p>
-          <div className="hero-ctas">
-            <a className="btn btn-primary" href={DOWNLOAD_URL}>
-              <Download aria-hidden />
-              Download for macOS
-            </a>
-            <a className="btn btn-outline" href={GITHUB_URL}>
-              View on GitHub
-            </a>
-          </div>
-          <BrewCmd />
-          <div className="hero-meta">Signed &amp; notarized · no telemetry · macOS 14+</div>
-        </div>
-      </section>
-
+      <HowItWorks />
+      <Features />
+      <Cli />
+      <Agents />
+      <Privacy />
+      <Faq />
+      <Cta />
       <LaunchBadgeTrack />
-
-      <footer>
-        <div className="container footer-inner">
-          <span className="muted">munkel · ephemeral messages between friends.</span>
-          <div className="footer-links">
-            <a href={GITHUB_URL}>GitHub</a>
-            <a href={PROTOCOL_URL}>Protocol v1</a>
-            <a href={CLI_URL}>munkel CLI</a>
-            <a href={LICENSE_URL}>MIT License</a>
-          </div>
-        </div>
-      </footer>
-    </>
+      <SiteFooter />
+    </MotionConfig>
   )
 }
