+++ apps/landing/vite.config.ts
@@ -1,3 +1,5 @@
+import { fileURLToPath } from 'node:url'
+
 import { defineConfig } from 'vite'
 import { devtools } from '@tanstack/devtools-vite'
 
@@ -7,8 +9,10 @@ import viteReact from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
 import { cloudflare } from '@cloudflare/vite-plugin'
 
+const srcDir = fileURLToPath(new URL('./src', import.meta.url))
+
 const config = defineConfig({
-  resolve: { tsconfigPaths: true },
+  resolve: { alias: { '@': srcDir } },
   plugins: [
     devtools(),
     cloudflare({ viteEnvironment: { name: 'ssr' } }),
